middleware checks, decorators checks .. etc

+++ Admin/views.py
@@ -4,8 +4,9 @@ import re
 import json
 from django.shortcuts import render, redirect
 from django.http import HttpRequest, Http404, HttpResponse, JsonResponse
-from .models import Admin, Role, Permission
+from .models import Admin, Role, Permission, Category, Order, Product, Client
 from .forms import CreateAdminForm, UpdateAdminForm, LoginForm
+from .decorators import is_allowed as is_allowed_decorator
 from django.views.decorators.http import require_http_methods
 from django.contrib import messages
 from django.core import serializers
@@ -18,14 +19,60 @@ from django.utils.translation import gettext as _
 from django.core.paginator import Paginator
 from django.db.models import Q
 
+
+
 def index_admin(request):
+    users_count = Admin.objects.count
+    categories_count = Category.objects.count
+    products_count    = Product.objects.count
+    clients_count    = Client.objects.count
+    orders_count      = Order.objects.count
+
+    data =  {
+        "users": {
+            "name": _("users"),
+            "href": "/admin/all/",
+            "count": users_count,
+            "icon": "ion ion-ios-people text-light",
+            "color": "primary",
+        },
+        "categories": {
+            "name": _("categories"),
+            "href": "/admin/category/all/",
+            "count": categories_count,
+            "icon": "ion ion-ios-list text-light",
+            "color": "success",
+        },
+        "products": {
+            "name": _("products"),
+            "href": "/admin/product/all/",
+            "count": products_count,
+            "icon": "ion ion-ios-box text-light",
+            "color": "info",
+        },
+        "clients": {
+            "name": _("clients"),
+            "href": "/admin/client/all/",
+            "count": clients_count,
+            "icon": "ion ion-ios-contacts text-light",
+            "color": "navy",
+        },
+        "orders": {
+            "name": _("orders"),
+            "href": "/admin/order/all/",
+            "count": orders_count,
+            "icon": "ion ion-md-paper-plane text-light",
+            "color": "secondary",
+        },
+    }
+    ctx = {
+        "data": data
+    }
     
-    # from .locale.ar import lang as _
-    # return HttpResponse(_["name"])
-    return render(request,"index.html");
+    return render(request,"index.html", ctx);
 
 
-
+@is_allowed_decorator("create_admin")
 def create_admin(request):
 
     METHOD = request.method
@@ -82,20 +129,20 @@ def create_admin(request):
             return redirect("/admin/")
 
 
-
 def edit_admin(request, id):
-    old_user = Admin.objects.prefetch_related('permission').get(id= id)
-    db_user = old_user
+    user = Admin.objects.prefetch_related('permission').get(id= id)
+    
     db_roles = Role.objects.all()
     db_permissions = Permission.objects.all()
 
     if request.method == "GET":
         user_permission_array = []
-        for i in db_user.permission.all():
+        for i in user.permission.all():
             user_permission_array.append(i.name)
 
+        
         return render(request, "administrators/edit.html", {
-            "user": db_user,
+            "user": user,
             "roles": db_roles,
             "permissions": db_permissions,
             "user_permission_array": user_permission_array
@@ -104,13 +151,13 @@ def edit_admin(request, id):
     elif request.method == "POST":
         user_permission_array = []
         
-        for i in db_user.permission.all():
+        for i in user.permission.all():
             user_permission_array.append(i.name)
 
-        context = {"user": db_user, "roles": db_roles, "permissions": db_permissions,  "user_permission_array": user_permission_array}
+        context = {"user": user, "roles": db_roles, "permissions": db_permissions,  "user_permission_array": user_permission_array}
         
         updateAdminForm = UpdateAdminForm(request.POST, request.FILES)
-        user = old_user
+        # user = old_user
         
         if not  updateAdminForm.is_valid():
             context ["form"] = updateAdminForm
@@ -176,7 +223,7 @@ def edit_admin(request, id):
             for s in Session.objects.all():
                 if s.get_decoded().get("user_id") == user.id and not is_own_profile(request, user.id):
                     s.delete()
-
+            
 
             messages.success(request, _("user updated successfully"))
             return redirect(request.META.get("HTTP_REFERER"))
@@ -185,33 +232,42 @@ def edit_admin(request, id):
 
 
 def detailed_admin(request, id):
-    if request.method == "GET":
-        
+    
+    if not is_allowed(request, id, "read_admin"):
+        if is_own_profile(request, id):
+            if request.method == "GET":
+                db_user = Admin.objects.prefetch_related('permission', "role").get(id= id)
+                return render(request, "administrators/detailed_user.html", {"user": db_user} )
+            else:
+                return redirect("/admin/error403")
+    else:
         db_user = Admin.objects.prefetch_related('permission', "role").get(id= id)
         return render(request, "administrators/detailed_user.html", {"user": db_user} )
-  
+        
 
 
-
-
-      
+@is_allowed_decorator("delete_admin")      
 def delete_admin(request, id):
     
     context = {}
     if request.method == "POST":
         
         user = Admin.objects.get(id = id)
-        if user.delete():
-            return JsonResponse({"success": True , "message": "product has been deleted successfully"})
+        if user.role.name != "super_admin":
+            if user.delete():
+                return JsonResponse({"success": True , "message": "user has been deleted successfully"})
 
+            else:
+                return JsonResponse({"success": False , "message": "can't delete this user"})
         else:
-            return JsonResponse({"success": False , "message": "can't delete this product"})
+            return JsonResponse({"success": False , "message": "can't delete this user"})
+            
     else:
         return JsonResponse({"error": "wrong method"})
 
 
     
-        
+@is_allowed_decorator("read_admin")        
 def all_admins(request):
     per_page = 2
     page = request.GET.get("page", 1)
